Change clockFreq from int to float
New Fast2D is 33.3333 instead of 33 Mhz.

+++ Probe.h
@@ -173,7 +173,7 @@ protected:
    * Clock speed of the probe.  Not used by original PMS2D, but used by most modern
    * probes to decode the timing word.
    */
-  unsigned int	_clockMhz;
+  float		_clockMhz;
 
   /**
    * Number of bins being output for size-distribution. Values are nDiodes for
